fix: Diary Alert 삭제 오류 수정
- 자기 자신이 쓴 댓글 또는 좋아요는 알림에 추가되지 않기 때문에 생긴 오류
- Diary 글쓴이와 알림의 글쓴이가 다르면 삭제 진행

+++ src/main/java/com/phcworld/service/board/DiaryServiceImpl.java
@@ -124,13 +124,17 @@ public class DiaryServiceImpl implements DiaryService {
 		for(int i = 0; i < goodList.size(); i++) {
 			Good good = goodList.get(i);
 //			timelineService.deleteTimeline(goodList.get(i));
-			alertService.deleteAlert(good);
+			if(!diary.getWriter().equals(good.getUser())){
+				alertService.deleteAlert(good);
+			}
 		}
 		List<DiaryAnswer> answerList = diary.getDiaryAnswers();
 		for(int i = 0; i < answerList.size(); i++) {
 			DiaryAnswer diaryAnswer = answerList.get(i);
 //			timelineService.deleteTimeline(diaryAnswer);
-			alertService.deleteAlert(diaryAnswer);
+			if(!diary.getWriter().equals(diaryAnswer.getWriter())){
+				alertService.deleteAlert(diaryAnswer);
+			}
 		}
 		diaryRepository.delete(diary);
 	}
